+++ pycololight/__init__.py
@@ -1,10 +1,15 @@
 from .pycololight import (
     PyCololight,
-    ColourSchemeException,
-    ColourException,
-    CycleSpeedException,
-    ModeExecption,
     DefaultEffectExecption,
     BrightnessException,
     UnavailableException,
+    UnsupportedDeviceException,
+)
+
+from .effects import (
+    Effects,
+    CycleSpeedException,
+    ColourSchemeException,
+    ColourException,
+    ModeExecption,
 )