+++ pycololight/pycololight.py
@@ -1,27 +1,10 @@
 import socket
 
+from pycololight.effects import Effects
+
 from .constants import (
     COMMAND_PREFIX,
-    CUSTOM_EFFECT_COLOURS,
-    CUSTOM_EFFECT_MODES,
-    DEFAULT_EFFECTS,
 )
-
-
-class ColourSchemeException(Exception):
-    pass
-
-
-class ColourException(Exception):
-    pass
-
-
-class CycleSpeedException(Exception):
-    pass
-
-
-class ModeExecption(Exception):
-    pass
 
 
 class DefaultEffectExecption(Exception):
@@ -36,12 +19,20 @@ class UnavailableException(Exception):
     pass
 
 
+class UnsupportedDeviceException(Exception):
+    pass
+
+
 class PyCololight:
     """
     Cololight wrapper
     """
 
-    def __init__(self, host, port=8900, default_effects=True):
+    def __init__(
+        self, device, host, port=8900, default_effects=True, dynamic_effects=False
+    ):
+        self.supported_devices = ["hexagon", "strip"]
+        self.device = self._check_supported_devices(device)
         self.host = host
         self.port = port
         self._counter = 1
@@ -49,8 +40,25 @@ class PyCololight:
         self._brightness = None
         self._colour = None
         self._effect = None
-        self._effects = DEFAULT_EFFECTS.copy() if default_effects else {}
+        self._device_effects = Effects(device)
+        self._effects = self._initial_effects(default_effects, dynamic_effects)
         self._sock = None
+
+    def _check_supported_devices(self, device):
+        if device not in self.supported_devices:
+            raise UnsupportedDeviceException
+        else:
+            return device
+
+    def _initial_effects(self, default_effects, dynamic_effects):
+        initial_effects = {}
+
+        if default_effects:
+            initial_effects.update(self._device_effects.default_effects)
+        if dynamic_effects:
+            initial_effects.update(self._device_effects.dynamic_effects)
+
+        return initial_effects
 
     def _toggle_counter(self):
         self._counter = 2 if self._counter == 1 else 1
@@ -67,9 +75,10 @@ class PyCololight:
     def _socket_close(self):
         self._sock.close()
 
-    def _send(self, command, response=False):
+    def _send(self, commands, response=False):
         self._socket_connect()
-        self._sock.sendto(command, (self.host, self.port))
+        for command in commands:
+            self._sock.sendto(command, (self.host, self.port))
 
         if not response:
             self._socket_close()
@@ -93,47 +102,13 @@ class PyCololight:
 
         return config
 
-    def _cycle_speed_hex(self, cycle_speed, mode):
-        if not 1 <= cycle_speed <= 32:
-            raise CycleSpeedException
-        if mode in [2]:
-            # Mode 2 only has speeds 1, 2, 3, which are mapped differently to other modes
-            cycle_speed_values = [3, 11, 19]
-            cycle_speed_value = cycle_speed_values[min(3, cycle_speed) - 1]
-        else:
-            cycle_speed_value = list(reversed(range(33)))[cycle_speed - 1]
-
-        cycle_speed_hex = "{:02x}".format(cycle_speed_value)
-        return cycle_speed_hex
-
-    def _colour_hex(self, colour_scheme, colour, mode):
-        if colour_scheme not in self.custom_effect_colour_schemes():
-            raise ColourSchemeException
-        if colour not in self.custom_effect_colour_scheme_colours(colour_scheme):
-            raise ColourException
-
-        starting_decimal = CUSTOM_EFFECT_COLOURS[colour_scheme]["decimal"]
-        colour_key = CUSTOM_EFFECT_COLOURS[colour_scheme]["colours"].index(colour)
-        if mode in [13, 14, 15, 22, 23, 24]:
-            # These modes have a lower starting decimal of 128
-            starting_decimal = starting_decimal - 128
-        colour_decimal = starting_decimal + colour_key
-        colour_hex = "{:02x}".format(colour_decimal)
-        return colour_hex
-
-    def _mode_hex(self, mode):
-        if not 1 <= mode <= len(CUSTOM_EFFECT_MODES):
-            raise ModeExecption
-
-        return CUSTOM_EFFECT_MODES[mode - 1]
-
     @property
     def state(self):
         """
         Gets state (on/off) and brightness from device, and updates local state.
         """
         self._send(
-            bytes.fromhex(f"{self._get_config('state')}"),
+            [bytes.fromhex(f"{self._get_config('state')}")],
             response=True,
         )
         data = self._receive()
@@ -163,7 +138,7 @@ class PyCololight:
         else:
             self._on = False
             command = bytes.fromhex("{}{}".format(self._get_config("command"), "e1e"))
-            self._send(command)
+            self._send([command])
 
     @property
     def brightness(self) -> int:
@@ -188,7 +163,7 @@ class PyCololight:
             )
         )
         self._brightness = brightness
-        self._send(command)
+        self._send([command])
 
     @property
     def colour(self) -> tuple:
@@ -209,7 +184,7 @@ class PyCololight:
             )
         )
         self._colour = colour
-        self._send(command)
+        self._send([command])
 
     @property
     def effect(self) -> str:
@@ -223,21 +198,32 @@ class PyCololight:
         """
         Set the effect of the device.
         """
-        command = bytes.fromhex(
-            "{}{}".format(
-                self._get_config("effect"),
-                self._effects[effect],
+        effect_commands = self._effects[effect]
+        commands = []
+        for effect_command in effect_commands:
+            command = bytes.fromhex(
+                "{}{}".format(
+                    self._get_config("effect"),
+                    effect_command,
+                )
             )
-        )
+            commands.append(command)
         self._effect = effect
-        self._send(command)
+        self._send(commands)
 
     @property
     def default_effects(self) -> list:
         """
         Returns a list of names of the default effects for the device.
         """
-        return list(DEFAULT_EFFECTS.keys())
+        return list(self._device_effects.default_effects.keys())
+
+    @property
+    def dynamic_effects(self) -> list:
+        """
+        Returns a list of names of the dynamic effects for the device.
+        """
+        return list(self._device_effects.dynamic_effects.keys())
 
     @property
     def effects(self) -> list:
@@ -246,15 +232,15 @@ class PyCololight:
         """
         return list(self._effects.keys())
 
-    def restore_default_effects(self, effects: list):
+    def restore_effects(self, effects: list):
         """
-        Restors the given default effects, to the saved effects of the device.
+        Restors the given default/dynamic effects, to the saved effects of the device.
         """
         for effect in effects:
-            if effect not in DEFAULT_EFFECTS:
+            if effect not in self._device_effects.effects.keys():
                 raise DefaultEffectExecption
 
-            self._effects[effect] = DEFAULT_EFFECTS[effect]
+            self._effects[effect] = self._device_effects.effects[effect]
 
     def add_custom_effect(
         self, name: str, colour_scheme: str, colour: str, cycle_speed: int, mode: int
@@ -262,30 +248,20 @@ class PyCololight:
         """
         Adds a custom effect, to the saved effects of the device.
         """
-        cycle_speed_hex = self._cycle_speed_hex(int(cycle_speed), int(mode))
-        colour_hex = self._colour_hex(colour_scheme, colour, int(mode))
-        mode_hex = self._mode_hex(int(mode))
+        custom_effect_hex = self._device_effects.custom_effect_command(
+            colour_scheme, colour, cycle_speed, mode
+        )
 
-        if mode in [2]:
-            # Mode 2 has bytes arranged differently to other modes
-            custom_effect_hex = (
-                f"{mode_hex[0]}{cycle_speed_hex}{colour_hex}{mode_hex[1]}"
-            )
-        else:
-            custom_effect_hex = (
-                f"{mode_hex[0]}{colour_hex}{cycle_speed_hex}{mode_hex[1]}"
-            )
-
-        self._effects[name] = custom_effect_hex
+        self._effects[name] = [custom_effect_hex]
 
     def custom_effect_colour_schemes(self) -> list:
         """
         Returns a list of the available colour schemes for custom efects.
         """
-        return list(CUSTOM_EFFECT_COLOURS.keys())
+        return self._device_effects.custom_effect_colour_schemes()
 
     def custom_effect_colour_scheme_colours(self, colour_scheme) -> list:
         """
         Returns a list of the available colours for a given colour schemes for custom efects.
         """
-        return list(filter(None, CUSTOM_EFFECT_COLOURS[colour_scheme]["colours"]))
+        return self._device_effects.custom_effect_colour_scheme_colours(colour_scheme)